Fix: LData 관련 메소드 합침, 테스트 수정
