+++ src/main/dataStructure/List/LData.java
@@ -1,6 +1,11 @@
 package main.dataStructure.List;
 
 public class LData {
+  public static int EQUAL_ALL = 0;
+  public static int EQUAL_X = 1;
+  public static int EQUAL_y = 2;
+  public static int UNEQUAL = -1;
+
   private int x;
   private int y;
 
@@ -30,5 +35,20 @@ public class LData {
   public void reset() {
     this.x = 0;
     this.y = 0;
+  }
+
+  static public int compareLData(LData x, LData y) {
+    if (x.getX() == y.getX() && x.getY() == y.getY()) {
+      return EQUAL_ALL;
+    } else if (x.getX() == y.getX()) {
+      return EQUAL_X;
+    } else if (x.getY() == y.getY()) {
+      return EQUAL_y;
+    }
+    return UNEQUAL;
+  }
+
+  static public void showData(LData lData) {
+    System.out.println("x: " + lData.getX() + ", y: " + lData.getY());
   }
 }
